Loosen the Result::Err bound
This crate is currently not ergonomic to use with something like anyhow.
Crates like anyhow can't implement `std::error::Error` for their error
types, because that would create multiple `From<anyhow::Error>`
implementations.

I originally wrote the trait bound as `std::error::Error` to make it
clear that should be, well, an error, but `Result<T, E>` already makes
that pretty clear.

The only things we do with the value are get the type name and convert
the type to a string, so I've loosened the bound to make the library
much more convenient to use.